Shock Results Obtained for Subsonic and Supersonic Conditions
The following results are calculations implemented from testing the
vectorized scheme with various flow regimes, domain sizes, and
grid-spacing. Currently, the geometry tested here is the cylinder flow.
This is to partially test the euler scheme for cylindrical flows as well
as to speed up the calculations from previous prototyped solvers. A
previous update included the initial runs for the Cartesian grid over
the biconvex, however it currently just uses raw serial programming to
ensure equations are properly implemented. A revisit to vectorize and
expand to different results will be necessary.

Results need to be confirmed with Professor Hafez before pushed to
master.