Dufort-Frankel Method Implemented for Full Vectorized Operations
A Dufort-Frankel-like method has been implemented in the equations to
assist convergence and automatically set the parameter for the
time-based dissipation factor/viscosity. The method dutilizes the
unconditional stability of the intrinsic integration method to overcome
stability issues with solving the heat equation. This is done by
substituting the values at the current node in the 2D space, (i,j), with
an averaged value of the adjacent time steps (forward and backward) in
the same location. The formulation can then be rewritten to an explicit
formulation to find that the scheme now only needs the spatial viscosity
term.

Past attempts to implement it resulted in very cluttered code. A
variable defined for the exchange of terms ("DF_coeff") simplified the
implementation process. This commit includes two main test-cases in a
supersonic (M=1.2) and subsonic (M=0.53) flow regimes.