Rewrote Solver to Improve Distinction Between Solver and Equations
The solver is now written in such a way that allows it to be implemented
as a separate entity interchangeable with other fluid system of
equations. This can be observed through the use of a separate function
to run the time-stepping schemes, another independent function that
calculates the boundary conditions, and a final independent function
that calculates the lower derivatives in the function. The purpose of
this is to allow for different time-stepping schemes to be implemented
smoothly, in case the Dufort-Frankel method is not sufficient for more
complicated systems.