Completed Testing of DF method on 2-Eqn Potential Flow
Dufort Frankel Pseudo-Time iteration has been successfully completed for
the two-equation potential flow system. The branch should now contain
the PDE solutions of high-subsonic to supersonic flows over an airfoil
and cylinder. The next step is to merge the various iterations of the
solver, equation models, and boundary condition solver.