Added New Solver to Solve Euler Eqns Explicitly as Hyperbolic System
Added a solver called threeLevelExplicit.m that solves the inviscid, compressible, flow equations explicitly as a hyperbolic system. It also  provides the framework for additional modular forces, diffusions, etc, to be added to the system if one so desires. The other solver, dufortFrankel.m, should no longer be used for solving the Euler equations. Instead it will be re-purposed for solving heat-like (parabolic) systems in the future. Initial progress on adding the JST scheme addition is also included. The vonNeumann-Richtmyer scheme has been modified for the new threeLevelExplicit framework.